pyproject.toml
[project]
name = "sherlock"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "instructor[bedrock]",
    "anthropic>=0.18.0",
    "networkx>=3.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pytest-dotenv>=0.5.2",
    "pytest>=8.3.3",
    "ruff>=0.7.4",
]
